pyproject.toml
[project]
name = "workshop-automation"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "requests",
    "python-dotenv",
    "openai>=2.15.0",
    "mastodon-py>=2.1.4",
    "replicate>=1.0.7",
]

[dependency-groups]
dev = []
